프로젝트 병합후 수정

+++ src/main/java/com/finalProject/togOther/domain/Planner.java
@@ -2,26 +2,20 @@ package com.finalProject.togOther.domain;
 
 import java.sql.Timestamp;
 
-import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.CreationTimestamp;
 
-import com.finalProject.togOther.constant.Continent;
 import com.finalProject.togOther.dto.PlannerDTO;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
@@ -33,7 +27,7 @@ public class Planner {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int plannerSeq;
-	private int code;
+	private byte code;
 	private String title;
 	private String startDate;
 	private String endDate;
@@ -42,22 +36,22 @@ public class Planner {
 	private Timestamp logTime;
 	private int hit;
 	private int likeCnt;
-	private int publicPlan;
+	private byte publicPlan;
 	
 	
 	public static Planner toEntity(PlannerDTO plannerDTO) {
 		
 		return Planner.builder()
-				.plannerSeq(plannerDTO.getPlannerSeq())
-				.code(plannerDTO.getCode())
-				.title(plannerDTO.getTitle())
-				.startDate(plannerDTO.getStartDate())
-				.endDate(plannerDTO.getEndDate())
-				.logTime(plannerDTO.getLogTime())
-				.hit(plannerDTO.getHit())
-				.likeCnt(plannerDTO.getLikeCnt())
-				.publicPlan(plannerDTO.getPublicPlan())
-				.build();
+					  .plannerSeq(plannerDTO.getPlannerSeq())
+					  .code(plannerDTO.getCode())
+					  .title(plannerDTO.getTitle())
+					  .startDate(plannerDTO.getStartDate())
+					  .endDate(plannerDTO.getEndDate())
+					  .logTime(plannerDTO.getLogTime())
+					  .hit(plannerDTO.getHit())
+					  .likeCnt(plannerDTO.getLikeCnt())
+					  .publicPlan(plannerDTO.getPublicPlan())
+					  .build();
 	}
 	
 }
