+++ src/main/java/com/finalProject/togOther/dto/PlannerDTO.java
@@ -20,29 +20,29 @@ import lombok.Setter;
 @AllArgsConstructor
 public class PlannerDTO {
 	private int plannerSeq;
-	private int code;
+	private byte code;
 	private String title;
 	private String startDate;
 	private String endDate;
 	private Timestamp logTime;
 	private int hit;
 	private int likeCnt;
-	private int publicPlan;
+	private byte publicPlan;
 	
 	
 	public static PlannerDTO toDTO(Planner planner) {
 		
 		return PlannerDTO.builder()
-				.plannerSeq(planner.getPlannerSeq())
-				.code(planner.getCode())
-				.title(planner.getTitle())
-				.startDate(planner.getStartDate())
-				.endDate(planner.getEndDate())
-				.logTime(planner.getLogTime())
-				.hit(planner.getHit())
-				.likeCnt(planner.getLikeCnt())
-				.publicPlan(planner.getPublicPlan())
-				.build();
+						 .plannerSeq(planner.getPlannerSeq())
+						 .code(planner.getCode())
+						 .title(planner.getTitle())
+						 .startDate(planner.getStartDate())
+						 .endDate(planner.getEndDate())
+						 .logTime(planner.getLogTime())
+						 .hit(planner.getHit())
+						 .likeCnt(planner.getLikeCnt())
+						 .publicPlan(planner.getPublicPlan())
+						 .build();
 	}
 	
 }
